lab1/requirements.txt
strands
boto3
python-dotenv
